src/requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Babel==0.9
Flask-Cors==2.1.2
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-PyMongo==0.3.1
Flask-RESTful==0.3.3
Flask-Script==2.0.5
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.10.4
aniso8601==1.0.0
blinker==1.3
itsdangerous==0.24
pymongo==2.8.1
pytz==2015.4
requests==2.7.0
six==1.9.0
speaklater==1.3
wheel==0.24.0
